this will be crucial
the thing that is originating so much complexity, is the fact that each proxy has its own reference frame of sorts, each proxy needs to magically keep track of the position of the particle, in its own ref frame. which is none sense. by storing the original position, I can do intersections starting from there. that way each proxy is keeping distances relative to the same origin.