task-42 Start mvp client base project.
After trial and error with integrating vtk.js into the webserver, I've found that I will have an easier time using TypeScript is easier.

Started the client code to see how it might go. I have a good feeling about it.

New tasks will be opened to work on the client side code. This branch will therefere be merged as soon as possible.